requirements.txt
certifi==2018.10.15
chardet==3.0.4
future==0.17.1
idna==2.7
numpy==1.15.2
Pillow==5.3.0
PyYAML==3.13
requests==2.20.0
six==1.11.0
torch==0.4.1
torchtext==0.3.1
torchvision==0.2.1
tqdm==4.28.1
urllib3==1.24
